When to extensive Don't Repeat Yourself goes wrong
The end result of this solution does not repeat lines of code but
ends of substantially longer than its copied equivalent.

While the use of std::bind and std::function is very powerful, this
simply costs to much in readability and maintainability. In addition,
the requirement to use std::placeholder_1 _2 _3 etc feels dirty and
clunky.

In the end I keep this on a branch just for demonstration purposes.